# api/src/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Mnemosyne API",
        "version": "0.1.0",
        "description": "Local file semantic search REST API"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Default local server"
        }
    ],
    "paths": {
        "/health": {
            "get": {
                "summary": "Health check",
                "tags": [
                    "meta"
                ],
                "responses": {
                    "200": {
                        "description": "Service is up",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "0.1.0"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stats": {
            "get": {
                "summary": "Index statistics",
                "tags": [
                    "index"
                ],
                "responses": {
                    "200": {
                        "description": "IndexStats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total_files": {
                                            "type": "integer"
                                        },
                                        "total_chunks": {
                                            "type": "integer"
                                        },
                                        "files_by_type": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/search": {
            "post": {
                "summary": "Semantic / keyword / hybrid search",
                "tags": [
                    "search"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "机器学习"
                                    },
                                    "mode": {
                                        "type": "string",
                                        "enum": [
                                            "Hybrid",
                                            "Vector",
                                            "Keyword"
                                        ],
                                        "default": "Hybrid"
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "default": 20
                                    },
                                    "vector_weight": {
                                        "type": "number",
                                        "default": 0.7
                                    },
                                    "keyword_weight": {
                                        "type": "number",
                                        "default": 0.3
                                    },
                                    "file_types": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of SearchResult",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SearchResult"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/files": {
            "get": {
                "summary": "List indexed files",
                "tags": [
                    "files"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 200
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of FileRecord",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FileRecord"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/files/{id}": {
            "delete": {
                "summary": "Remove file from index",
                "tags": [
                    "files"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File removed"
                    }
                }
            }
        },
        "/api/index": {
            "post": {
                "summary": "Index a directory recursively",
                "tags": [
                    "index"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "path"
                                ],
                                "properties": {
                                    "path": {
                                        "type": "string",
                                        "example": "/Users/you/Documents"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "IndexStats"
                    }
                }
            }
        },
        "/api/models": {
            "get": {
                "summary": "List available / downloaded models",
                "tags": [
                    "models"
                ],
                "responses": {
                    "200": {
                        "description": "Model list"
                    }
                }
            }
        },
        "/api/models/download": {
            "post": {
                "summary": "Download a model from HuggingFace Hub",
                "tags": [
                    "models"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "model_id"
                                ],
                                "properties": {
                                    "model_id": {
                                        "type": "string",
                                        "example": "sentence-transformers/all-MiniLM-L6-v2"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Download started"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "FileRecord": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "path": {
                        "type": "string"
                    },
                    "file_type": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer"
                    }
                }
            },
            "SearchResult": {
                "type": "object",
                "properties": {
                    "file_record": {
                        "$ref": "#/components/schemas/FileRecord"
                    },
                    "score": {
                        "type": "number"
                    },
                    "snippet": {
                        "type": "string"
                    },
                    "match_type": {
                        "type": "string"
                    },
                    "chunk_index": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}